+++ export_test.go
@@ -1,4 +1,4 @@
-package langfuse
+package lunte
 
 // SDKVersion exposes the internal version constant so external wire tests can
 // assert the exact x-langfuse-sdk-version header value.