+++ internal/attributes/attributes_test.go
@@ -13,7 +13,7 @@ import (
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 
-	lfattr "github.com/fgn/langfuse-go/internal/attributes"
+	lfattr "github.com/fgn/lunte/internal/attributes"
 )
 
 type pointerOnlyLargeString string