+++ lifecycle_adversary_test.go
@@ -1,4 +1,4 @@
-package langfuse
+package lunte
 
 import (
 	"context"
@@ -16,8 +16,8 @@ import (
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	commonpb "go.opentelemetry.io/proto/otlp/common/v1"
 
-	lfattr "github.com/fgn/langfuse-go/internal/attributes"
-	"github.com/fgn/langfuse-go/internal/otlpreceiver"
+	lfattr "github.com/fgn/lunte/internal/attributes"
+	"github.com/fgn/lunte/internal/otlpreceiver"
 )
 
 func TestBorrowedProviderOnStartCanReenterClientShutdown(t *testing.T) {